requirements.txt
anyio==4.3.0
certifi==2024.2.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
python-telegram-bot==21.1.1
sniffio==1.3.1
typing_extensions==4.11.0
